adding rsc_check to Chef Install, and being explicit in haproxy

+++ chef-client/Chef_Client_Install.sh
@@ -110,6 +110,10 @@ if [[ $(dmidecode | grep -i 'Microsoft Corporation') ]] ; then
 EOF
 fi
 
+if [ ! -e /usr/local/bin/rsc ]; then
+  echo "rsc not found, RL10 is a requirement for the chef10 scripts"
+  exit 1
+fi
 
 cat > $chef_dir/client.rb <<-EOF
 log_level              $LOG_LEVEL
